+++ src/app/[lng]/learn/page.tsx
@@ -1,10 +1,6 @@
-import dynamic from 'next/dynamic';
 import { getArticles } from 'src/app/lib/getArticles';
 import { getFeaturedArticle } from 'src/app/lib/getFeaturedArticle';
-
-const LearnPage = dynamic(() => import('../../ui/learn/LearnPage'), {
-  ssr: true,
-});
+import LearnPage from 'src/app/ui/learn/LearnPage';
 
 // `app/ui/learn/page.tsx` is the UI for the `/learn` URL
 export default async function Page({
