+++ src/app/ui/app/App.tsx
@@ -1,4 +1,3 @@
-'use client';
 import { FeatureCards } from '@/components/FeatureCards/FeatureCards';
 import { Snackbar } from '@/components/Snackbar/Snackbar';
 import { WelcomeScreen } from '@/components/WelcomeScreen/WelcomeScreen';
