+++ src/providers/WalletProvider/WalletProvider.tsx
@@ -1,4 +1,3 @@
-'use client';
 import { type FC, type PropsWithChildren } from 'react';
 import { EVMProvider } from './EVMProvider';
 import { SolanaProvider } from './SolanaProvider';
